Fixes #319: Wrong username in Jenkins build description on push event
The build description is expected to provide information
about the user who triggered a build job.
Previously, the therein specified user was the committer who
most recently committed to that pushed branch. However,
this does not necessarily correspond to the GitLab user
who actually triggered the build job by pushing the changeset.
The existing behaviour was replaced to use the GitLab user
instead unless provided through the GitLab request and to
fall back to the most recent committer if the GitLab user
is not available.